# internal/openapi/spec/public.yaml
openapi: 3.0.3
info:
  title: metareel Public API
  version: "1.0.0"
  description: >
    Every /api/v1 endpoint needs an API key, sent as `X-API-Key: <key>` or
    `Authorization: Bearer <key>` (unless keys are turned off in the admin
    console). Keys are created and rotated in the admin console under
    Settings & API keys. Requests are rate limited per key; over the limit
    the API returns 429 with Retry-After. Missing or invalid keys get 401.
servers:
  - url: /
security:
  - ApiKeyHeader: []
  - BearerKey: []
paths:
...
  /api/v1/top10/{country}:
    get:
      operationId: getCountryTop10
      summary: Every provider's movie and TV Top 10 of a country in one response
      parameters:
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Charts (each provider's latest chart unless date is given)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Top10Combined"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/top10/global/{provider}:
    get:
      operationId: getGlobalTop10
      summary: A provider's titles ranked across every country by chart points
      description: >
        Points are FlixPatrol-style (rank 1 = 10 points … rank 10 = 1) summed
        over every stored country chart of the day. Titles mapped to the same
        TMDB title are one entry.
      parameters:
        - $ref: "#/components/parameters/ProviderParam"
        - $ref: "#/components/parameters/DateParam"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Leaderboard of one day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/movers:
    get:
      operationId: getMovers
      summary: Climbers, fallers, debuts, re-entries and exits of a chart
      description: >
        Compares the chart on `date` (default: the latest stored date) with
        the stored date before it. Debuts are titles not on the chart in the
        previous 90 days.
      parameters:
        - name: country
          in: query
          required: true
          schema: { type: string }
        - name: provider
          in: query
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/DateParam"
      responses:
        "200":
          description: Movers per category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoversResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/leaderboards:
    get:
      operationId: getLeaderboard
      summary: Titles ranked by chart points over a period
      description: >
        Points per chart per day: rank 1 = 10 … rank 10 = 1. Omit country /
        provider to sum over all. period=week (default) is the 7 days ending
        on `date` (default: the current chart date); custom uses from/to (at
        most 366 days).
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [day, week, month, custom], default: week }
        - $ref: "#/components/parameters/DateParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - name: provider
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/export/rankings:
    get:
      operationId: exportRankings
      summary: Chart entries over a range as CSV or NDJSON
      description: Defaults to the last 30 days; at most 366 days.
      parameters:
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - name: provider
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/CategoryQuery"
        - name: format
          in: query
          schema: { type: string, enum: [csv, ndjson], default: csv }
      responses:
        "200":
          description: >
            Columns / fields: date, country, provider, category, rank,
            season_number, title_id, slug, name, kind, tmdb_id, imdb_id
          content:
            text/csv: {}
            application/x-ndjson: {}
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/titles/lookup:
    get:
      operationId: lookupTitle
      summary: Find a title by IMDb ID, TMDB ID or FlixPatrol slug
      parameters:
        - name: imdb
          in: query
          schema: { type: string, example: tt12300742 }
        - name: tmdb
          in: query
          schema: { type: string, example: "movie:701387" }
        - name: slug
          in: query
          schema: { type: string, example: bugonia }
      description: Give exactly one of imdb, tmdb or slug.
      responses:
        "200":
          description: TMDB titles and FlixPatrol titles
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        kind: { type: string, enum: [movie, tv] }
                        tmdb_id: { type: string }
                  flixpatrol_titles:
                    type: array
                    items: { $ref: "#/components/schemas/RankedTitle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/titles/tmdb/ratings:
    get:
      operationId: getTitleRatingsBatch
      summary: Ratings of several titles, looked up by TMDB ID
      description: >
        Stored ratings within RATINGS_TTL are returned as is; up to 10 expired
        or unrated titles per request are refreshed from the providers. Titles
        still without ratings are listed in missing (the ratings pre-warm job
        keeps charting titles fresh).
      parameters:
        - name: ids
          in: query
          required: true
          schema: { type: string, example: "movie:425,tv:154385", description: At most 50 kind:tmdb_id pairs }
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/RatingsResponse" }
                  missing: { type: array, items: { type: string } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/titles/tmdb/{kind}/{id}:
    get:
      operationId: getTitleOverview
      summary: Everything about a title in one call
      description: >
        TMDB metadata, chart stats, current chart positions, ratings, where to
        watch and Netflix's official numbers. `include` picks sections
        (default: all). Ratings may be refreshed from the providers if expired.
        Sections that fail are named in `errors`; the rest are still returned.
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - $ref: "#/components/parameters/TmdbIDParam"
        - name: include
          in: query
          schema: { type: string, example: "metadata,stats,current,ratings,availability,netflix" }
        - name: country
          in: query
          schema: { type: string, description: Limit availability and current positions to a country }
      responses:
        "200":
          description: Title overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleOverview"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/titles/tmdb/{kind}/{id}/stats:
    get:
      operationId: getTitleStats
      summary: Chart performance of a title
      description: >
        Points, days on charts, peak rank, days at #1, debut, longest streak
        and how it spread across countries.
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - $ref: "#/components/parameters/TmdbIDParam"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleStats"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/titles/tmdb/{kind}/{id}/availability:
    get:
      operationId: getTitleAvailability
      summary: Where a title can be watched, per country (TMDB, data by JustWatch)
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - $ref: "#/components/parameters/TmdbIDParam"
        - name: country
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Offers per country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: TMDB metadata not fetched yet (needs TMDB_API_KEY and the metadata job)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/{kind}/{id}/netflix:
    get:
      operationId: getTitleNetflix
      summary: A title's weeks in Netflix's official Top 10
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - $ref: "#/components/parameters/TmdbIDParam"
      responses:
        "200":
          description: Global and per-country weeks
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind: { type: string }
                  tmdb_id: { type: string }
                  weeks_in_global_top10: { type: integer }
                  best_global_rank: { type: integer, nullable: true }
                  total_views: { type: integer }
                  total_hours_viewed: { type: integer }
                  countries: { type: array, items: { type: string } }
                  global: { type: array, items: { $ref: "#/components/schemas/NetflixWeekRow" } }
                  country_weeks: { type: array, items: { $ref: "#/components/schemas/NetflixWeekRow" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/netflix/most-popular:
    get:
      operationId: getNetflixMostPopular
      summary: Netflix's all-time most popular lists (views in the first 91 days)
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        category: { type: string }
                        rank: { type: integer }
                        show_title: { type: string }
                        season_title: { type: string, nullable: true }
                        hours_viewed_first_91_days: { type: integer, nullable: true }
                        runtime_hours: { type: number, nullable: true }
                        views_first_91_days: { type: integer, nullable: true }
                        kind: { type: string }
                        tmdb_id: { type: string, nullable: true }
                        imdb_id: { type: string, nullable: true }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/analytics/decay:
    get:
      operationId: getDecay
      summary: How titles fade after debuting on a chart
      description: >
        Follows every chart debut in the range (default: last 60 days) for 30
        days: retention (share still charting) and average rank per day since
        debut. Days whose chart was not stored are not counted.
      parameters:
        - $ref: "#/components/parameters/AnalyticsFrom"
        - $ref: "#/components/parameters/AnalyticsTo"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/ProviderQuery"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
      responses:
        "200":
          description: Decay curve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResult"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/analytics/country-similarity:
    get:
      operationId: getCountrySimilarity
      summary: Which countries watch alike (overlap of charting titles)
      description: Jaccard overlap of each pair of countries' charting titles in the range (default 30 days).
      parameters:
        - $ref: "#/components/parameters/AnalyticsFrom"
        - $ref: "#/components/parameters/AnalyticsTo"
        - $ref: "#/components/parameters/ProviderQuery"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
      responses:
        "200":
          description: Country pairs, most similar first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResult"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/analytics/release-lag:
    get:
      operationId: getReleaseLag
      summary: Days from release to first chart appearance
      description: >
        Uses the title's digital release date in the chart's country from TMDB
        when known, else TMDB's release / first air date. Needs TMDB metadata
        (titles.metadata job). Range default: last 90 days.
      parameters:
        - $ref: "#/components/parameters/AnalyticsFrom"
        - $ref: "#/components/parameters/AnalyticsTo"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/ProviderQuery"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
      responses:
        "200":
          description: Per-title lags with median / quartiles overall and per provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResult"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/analytics/ratings-vs-popularity:
    get:
      operationId: getRatingsVsPopularity
      summary: Chart points against IMDb and Rotten Tomatoes ratings
      description: >
        Pearson correlation of points with IMDb rating (IMDb dataset first,
        then stored provider ratings) and RT critics score, plus acclaimed but
        under-watched (IMDb ≥ 7.5, points below median) and popular but poorly
        rated (top-quartile points, IMDb < 6) lists. Only stored ratings are
        used. Range default: last 30 days.
      parameters:
        - $ref: "#/components/parameters/AnalyticsFrom"
        - $ref: "#/components/parameters/AnalyticsTo"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/ProviderQuery"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
      responses:
        "200":
          description: Titles with points and ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResult"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/analytics/genres:
    get:
      operationId: getGenres
      summary: Chart points by TMDB genre
      parameters:
        - $ref: "#/components/parameters/AnalyticsFrom"
        - $ref: "#/components/parameters/AnalyticsTo"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/ProviderQuery"
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/KindQuery"
      responses:
        "200":
          description: Genres by points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResult"
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/analytics/netflix-calibration:
    get:
      operationId: getNetflixCalibration
      summary: Netflix's official weekly views against FlixPatrol Netflix chart points
      description: >
        For each Netflix global Top 10 entry, the FlixPatrol Netflix chart
        points the same title earned that week (in `country`, or summed over
        countries), with Pearson r and the median views per point. Range
        default: last 84 days.
      parameters:
        - $ref: "#/components/parameters/AnalyticsFrom"
        - $ref: "#/components/parameters/AnalyticsTo"
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/KindQuery"
      responses:
        "200":
          description: Calibration rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResult"
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerKey:
      type: http
      scheme: bearer
      description: 'The same API key, as "Authorization: Bearer <key>"'
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
...
    CategoryQuery:
      name: category
      in: query
      required: false
      schema:
        type: string
        enum: [movies, tv_shows]
        description: Omit for both
...
    CountryParam:
      name: country
      in: path
      required: true
      schema:
        type: string
        description: ISO country code (e.g. US)
    ProviderParam:
      name: provider
      in: path
      required: true
      schema:
        type: string
        description: Provider slug
    DateParam:
      name: date
      in: query
      required: false
      schema:
        type: string
        format: date
        description: >
          Chart date in YYYY-MM-DD. Omit for the latest scraped chart; on the
          all-providers endpoints that is each provider's own latest chart
          (see the per-item date).
...
  schemas:
...
    Top10Combined:
      type: object
      properties:
        country: { type: string }
        movies: { type: array, items: { $ref: "#/components/schemas/Top10Result" } }
        tv_shows: { type: array, items: { $ref: "#/components/schemas/Top10Result" } }
        dates:
          type: object
          additionalProperties: { type: string, format: date }
          description: Chart date served per provider
...